meeting-recorder/requirements.txt
# UI Dependencies
tkinter
customtkinter>=5.2.0
Pillow>=10.0.0

# Audio Processing
pyaudio>=0.2.11
wave
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
pathlib
datetime

# Development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0